requirements.txt
numpy==1.24.1
pandas==1.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
